Move useEditTheme to EventTypeAndThemeStep
